Deleted .idea folder
The reason why this commit is there is because even though you may have added the folder to the .gitignore, the files will still remain in the repo.
When the repo is cloned, the .idea files will be downloaded. The only effect the gitignore will have is preventing any changes from showing up in git